backend/requirements.txt
django
django-environ
djangorestframework
django-filter
psycopg[binary]
drf-spectacular 
drf-spectacular-sidecar